Evaluate answer on keyup instead of keypress
By evaluating answer on *keyup* instead of *keypress* event, iOS will treat correct *fill in the blanks* answer, as correct before hitting *return* or *space*. This is also discussed on Stackoverflow [here](https://stackoverflow.com/questions/28944171/e-keycode-in-ios-safari-on-keyup-event). The *keypress* seems to be triggered before the last key pressed is added to the input value.