Removed tabindex from non-target Gap Fill texts
One could argue that also the non-target texts should have a tabindex, but I think its better to go from the answer target straight to the next answer target or the next button. 

Note: *Introductory Text* has `tabindex=1`, so start from 2. This was and still is the case after this commit.